Moved "About and support" to the settings fragment
