--- app/src/main/java/com/example/carpark/views/Support.java
@@ -1,51 +1,0 @@
-package com.example.carpark.views;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-
-import com.example.carpark.R;
-
-public class Support extends AppCompatActivity {
-
-    public TextView mEmail;
-    public EditText mSubject;
-    public EditText mMessage;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_support);
-
-        mEmail = (TextView) findViewById(R.id.mailID);
-        mMessage = (EditText)findViewById(R.id.messageID);
-        mSubject = (EditText)findViewById(R.id.subjectID);
-
-        Button fab = findViewById(R.id.send);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                sendMail();
-            }
-        });
-    }
-
-    private void sendMail() {
-
-        String mail = mEmail.getText().toString().trim();
-        String message = mMessage.getText().toString();
-        String subject = mSubject.getText().toString().trim();
-
-        //Send Mail
-        JavaMailAPI javaMailAPI = new JavaMailAPI(this,mail,subject,message);
-
-        javaMailAPI.execute();
-
-    }
-
-
-}
