+++ app/src/main/res/menu/nav_menu_items.xml
@@ -28,16 +28,6 @@
             android:title="My Vehicle" />
 
         <item
-            android:id="@+id/nav_about"
-            android:icon="@drawable/ic_about"
-            android:title="About" />
-
-        <item
-            android:id="@+id/nav_support"
-            android:icon="@drawable/ic_support"
-            android:title="Support" />
-
-        <item
             android:id="@+id/nav_settings"
             android:icon="@drawable/ic_settings"
             android:title="Settings" />
